refactored managers to services
