+++ app/Http/Controllers/Api/FeedsApiController.php
@@ -2,9 +2,12 @@
 
 namespace App\Http\Controllers\Api;
 
+use App\Contracts\ArticleContract;
 use App\Contracts\FeedContract;
 use App\Contracts\UserContract;
 use App\DTO\FeedFinder;
+use App\Jobs\ProcessFeeds;
+use App\Models\Feed;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Controller;
@@ -16,15 +19,18 @@ class FeedsApiController extends Controller
 {
     private FeedContract $feedDatabaseContract;
     private UserContract $userContract;
+    private ArticleContract $articleContract;
 
 
 
     public function __construct(
         FeedContract $feedDatabaseContract,
-        UserContract $userContract
+        UserContract $userContract,
+        ArticleContract $articleContract
     ) {
         $this->feedDatabaseContract = $feedDatabaseContract;
         $this->userContract = $userContract;
+        $this->articleContract = $articleContract;
     }
 
     public function index(): JsonResponse
@@ -58,9 +64,18 @@ class FeedsApiController extends Controller
         return response()->json($data);
     }
 
-    public function save(Request $request)
+    public function save(Request $request): JsonResponse
     {
         $data = $request->toArray();
-        dd($data);
+        $exist = Feed::where('url', $data['url'])->exists();
+        if(!$exist) {
+            $feed = new Feed();
+            $feed->fill($data);
+            $feed->save();
+            dispatch(new ProcessFeeds($feed->id, $this->articleContract));
+            return response()->json($feed);
+
+        }
+        return response()->json(['error' => 'Feed already exists']);
     }
 }
